+++ xenserver/resource_vbd.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/hashicorp/terraform/helper/hashcode"
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/ringods/go-xen-api-client"
+	"github.com/terra-farm/go-xen-api-client"
 )
 
 const (