+++ xenserver/types.go
@@ -23,7 +23,7 @@ import (
 	"log"
 	"strconv"
 
-	"github.com/ringods/go-xen-api-client"
+	"github.com/terra-farm/go-xen-api-client"
 )
 
 type Range struct {