+++ CHANGELOG.md
@@ -34,4 +34,4 @@ FEATURES:
 IMPROVEMENTS:
 
 * Set up Github releases via Travis CI.
-* Support for XenServer 7.2 via [`go-xen-api-client` #180cc7b](https://github.com/ringods/go-xen-api-client/commit/180cc7bfb7590fbc1a81c198b0011429ac58881f)
+* Support for XenServer 7.2 via [`go-xen-api-client` #180cc7b](https://github.com/terra-farm/go-xen-api-client/commit/180cc7bfb7590fbc1a81c198b0011429ac58881f)